Update libneo4j-omni version check to expect 5.0.7
This warning is meant to be triggered in the following cases:
- The user upgrades Neo4j::Bolt without having upgraded Neo4j::Client first.
- The user manually installs non-matching versions of Neo4j::Bolt and Neo4j::Client.

min_lib_version should always be set to the latest released library version.

The warning should be non-invasive.

For example, users might intentionally run non-matching versions when a newer library version contains a bug. Until that bug is fixed and a new version of Neo4j::Client has been made available, users shouldn't be unnecessarily bothered by warnings. The warning was so far controllable by disabling warnings in the "misc" category. But that category is used for a bunch of unrelated things in Perl, and selectively disabling it at load time might be a bit tricky, requiring users to fiddle with BEGIN etc.

With this change, the warning category becomes "Neo4j::Bolt".