Have automated tests use Neo4j::Client trial releases
If there is a newer trial release on Neo4j::Client available on CPAN, we'll usually like to test with that one instead of the stable release version. Using the --dev flag on cpanm accomplishes this. We're only doing this for Neo4j::Client, because we don't want trial releases of any other modules.

Note that the --dev flag is experimental; cpanm might not always find the latest trial release.

Additionally, this change updates the software versions used by the tests.